fix number type MonthlyChargesEnterpriseDrm change int to float32

+++ infrastructure/bunny_request_rest/create_video_library_response.go
@@ -85,5 +85,5 @@ type CreateVideoLibraryResponse struct {
 	PremiumEncodingPriceOverride            *int             `json:"PremiumEncodingPriceOverride"`
 	MonthlyChargesTranscribing              float32          `json:"MonthlyChargesTranscribing"`
 	MonthlyChargesPremiumEncoding           float32          `json:"MonthlyChargesPremiumEncoding"`
-	MonthlyChargesEnterpriseDrm             int              `json:"MonthlyChargesEnterpriseDrm"`
+	MonthlyChargesEnterpriseDrm             float32          `json:"MonthlyChargesEnterpriseDrm"`
 }
